4) added access modifiers, improved tests, removed casting

+++ src/main/java/core/basesyntax/Calculator.java
@@ -2,19 +2,18 @@ package core.basesyntax;
 
 public class Calculator {
     public double calculate(double first, char operation, double second) {
-
         switch (operation) {
             case '+':
-                return (double) first + second;
+                return first + second;
             case '-':
-                return (double) first - second;
+                return first - second;
             case '*':
-                return first == 0 || second == 0 ? 0 : (double) first * second;
+                return first == 0 || second == 0 ? 0 : first * second;
             case '/':
                 if (second == 0) {
                     throw new ArithmeticException("Sorry, I can't divide on zero");
                 }
-                return (double) first / second;
+                return first / second;
             case '^':
                 if (first == 0 && second < 0) {
                     throw new ArithmeticException("Sorry, I can't raising zero in negative pow");
